chore: plow through evrey api route with safe guards

+++ src/app/api/v0/document/delete/route.ts
@@ -1,17 +1,23 @@
+import { getCurrentUser } from '@/server/getCurrentUser';
 import { failResponse, okResponse } from '@/utils/response'
 import { supabase } from '@/utils/supabase'
 
 export async function DELETE(request: Request) {
-  const requestData = await request.json()
-  const { error } = await supabase
-    .from('posts')
-    .delete()
-    .match({ id: requestData.id })
+	const { error: userError } = await getCurrentUser();
+	if (userError) {
+		return failResponse('Trouble getting user')
+	}
 
-  if (error) {
-    console.error(error)
-    return failResponse(error?.message)
-  }
+	const requestData = await request.json()
+	const { error } = await supabase
+		.from('posts')
+		.delete()
+		.match({ id: requestData.id })
 
-  return okResponse('Document deleted')
+	if (error) {
+		console.error(error)
+		return failResponse(error?.message)
+	}
+
+	return okResponse('Document deleted')
 }
